requirements.txt
Mastodon.py==1.3.1